Move ValueRef forward declarations into place of use
