+++ msvc2017/FreeOrion/FreeOrion.vcxproj.filters
@@ -354,9 +354,6 @@
     <ClInclude Include="..\..\UI\PlayerListWnd.h">
       <Filter>Header Files\UI</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\universe\ValueRefFwd.h">
-      <Filter>Header Files\universe</Filter>
-    </ClInclude>
     <ClInclude Include="..\..\UI\TechTreeLayout.h">
       <Filter>Header Files\UI</Filter>
     </ClInclude>
