+++ msvc2017/FreeOrionCA/FreeOrionCA.vcxproj.filters
@@ -227,9 +227,6 @@
     <ClInclude Include="..\..\GG\GG\Clr.h">
       <Filter>Header Files\GG</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\universe\ValueRefFwd.h">
-      <Filter>Header Files\universe</Filter>
-    </ClInclude>
     <ClInclude Include="..\..\universe\ObjectMap.h">
       <Filter>Header Files\universe</Filter>
     </ClInclude>
