+++ parse/EnumParser.h
@@ -4,7 +4,7 @@
 #include "Lexer.h"
 
 #include "ParseImpl.h"
-#include "../universe/ValueRefFwd.h"
+#include "../universe/ValueRef.h"
 #include "../universe/EnumsFwd.h"
 
 #include <boost/spirit/include/qi.hpp>
